Add files via upload
Hi Todd and Kerry, think all files associated with the package managed to be uploaded together. The description needs to be updated, as well as a few other obvious changes. Didn't execute exactly like the book, although I'm not sure if that's a good or a bad thing for this assignment. Main thing is that the package will still be flawed as long as the main code isn't perfect. I'm going to get some rest, but I definitely intend to spend more time working on both the package and main code today, so we have time to ask Dr. Molinari about anything that's still giving us trouble.  Thanks again team.